+++ app.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 import motor
 import os
-import time
+import led 
 from importlib import import_module
 
 from flask import Flask, render_template, Response, request
@@ -19,13 +19,14 @@ else:
 app = Flask(__name__)
 app.secret_key = "vth"
 
-LEFT, RIGHT, FORWARD, BACKWARD, STOP = "left", "right", "forward", "backward", "stop"
+LEFT, RIGHT, FORWARD, BACKWARD, STOP,  DISTANCE = "left", "right", "forward", "backward", "stop", "distance"
 AVAILABLE_COMMANDS = {
     'Left': LEFT,
     'Forward': FORWARD,
     'Right': RIGHT,
     'Backward': BACKWARD,
-    'Stop': STOP
+    'Stop': STOP,
+    'Distance': DISTANCE
 }
 
 
@@ -53,113 +54,30 @@ def video_feed():
                     mimetype='multipart/x-mixed-replace; boundary=frame')
 
 
-# GPIO Mode
-GPIO.setmode(GPIO.BCM)
-# Set GPIO Pins
-LED_L = 2
-LED_R = 4
-# Set GPIO Pins
-TRIG = 18
-ECHO = 24
-LED = 23
-LED_S = 3
-GPIO.setup(TRIG, GPIO.OUT)
-GPIO.setup(ECHO, GPIO.IN)
-# LED setup
-GPIO.setwarnings(False)
-GPIO.setup(LED_L, GPIO.OUT)
-GPIO.setup(LED_R, GPIO.OUT)
-GPIO.setup(LED, GPIO.OUT)
-GPIO.setup(LED_S, GPIO.OUT)
-
-
-def led():
-    GPIO.output(LED_L, GPIO.LOW)
-    GPIO.output(LED_R, GPIO.LOW)
-    GPIO.output(LED_S, GPIO.LOW)
-
-
-def led_l():
-    GPIO.output(LED_L, GPIO.HIGH)
-    GPIO.output(LED_R, GPIO.LOW)
-    GPIO.output(LED_S, GPIO.LOW)
-
-
-def led_r():
-    GPIO.output(LED_R, GPIO.HIGH)
-    GPIO.output(LED_L, GPIO.LOW)
-    GPIO.output(LED_S, GPIO.LOW)
-
-
-def led_s():
-    GPIO.output(LED_R, GPIO.LOW)
-    GPIO.output(LED_L, GPIO.LOW)
-    GPIO.output(LED_S, GPIO.HIGH)
-
-
-def distance():
-    GPIO.output(TRIG, False)
-    time.sleep(0.1)
-    print("Measuring...")
-    GPIO.output(TRIG, True)
-    time.sleep(0.00001)
-    GPIO.output(TRIG, False)
-    start = time.time()
-    stop = time.time()
-
-    # Save StartTime
-    while GPIO.input(ECHO) == 0:
-        start = time.time()
-
-    # Save StopTime
-    while GPIO.input(ECHO) == 1:
-        stop = time.time()
-    return (stop - start) / 0.000058  # cm
-
 
 @app.route('/<cmd>')
 def command(cmd=None):
     if cmd == STOP:
-        print("Stop Pressed")
-        led_s()
-        measure()
+        led.led_stop()
         motor.stop()
     elif cmd == FORWARD:
-        print("Up Pressed")
-        led()
-        measure()
+        led.led_off()
         motor.forward()
     elif cmd == BACKWARD:
-        print("Down Pressed")
-        led()
-        measure()
+        led.led_off()
         motor.backward()
     elif cmd == LEFT:
-        print("Left Pressed")
-        led_l()
-        measure()
+        led.led_left()
         motor.left()
-    else:
-        print("Right Pressed")
-        led_r()
-        measure()
+    elif cmd == RIGHT:
+        led.led_right()
         motor.right()
+    if cmd == DISTANCE:
+        led.measure()
+            
     response = "Moving {}".format(cmd.capitalize())
     return response, 200, {'Content-Type': 'text/plain'}
 
-
-def measure():
-    try:
-        while True:
-            dist = distance()
-            print("Distance: %.2f cm" % dist)
-            if dist < 15:
-                GPIO.output(LED, GPIO.HIGH)
-            else:
-                GPIO.output(LED, GPIO.LOW)
-            time.sleep(1)
-    except KeyboardInterrupt:
-        print("Stop measurement")
 
 
 @app.route('/log_out')
