+++ templates/index.html
@@ -42,8 +42,11 @@
             <button class="down" onclick="moveDown()"><span>&#8595;</span></button>
         </p>
         <br><br>
+        <p>
+            <input type="button" class="off" onclick="distance()" value="Measure Distance">
+        </p>
+        <br><br>
         <input type="button" class="off" value="Turn Off" onclick="window.location.href='/log_out'"/>
-
         <br/><br/>
 
     </div>
